+++ core/src/commands/mod.rs
@@ -36,6 +36,7 @@ pub mod cmd_log;
 pub mod cmd_ls;
 #[cfg(all(feature = "fuse", target_os = "linux"))]
 pub mod cmd_mount;
+pub mod cmd_rebuild_index;
 pub mod cmd_recall;
 pub mod cmd_rechunk;
 pub mod cmd_restore;
@@ -72,6 +73,7 @@ pub enum Command {
     Ls(WithGlobal<cmd_ls::CmdArgs>),
     #[cfg(all(feature = "fuse", target_os = "linux"))]
     Mount(WithGlobal<cmd_mount::CmdArgs>),
+    RebuildIndex(WithGlobal<cmd_rebuild_index::CmdArgs>),
     Recall(WithGlobal<cmd_recall::CmdArgs>),
     Rechunk(WithGlobal<cmd_rechunk::CmdArgs>),
     Restore(WithGlobal<cmd_restore::CmdArgs>),
@@ -233,6 +235,7 @@ pub fn parse_and_run() -> Result<()> {
         Command::Ls(cmd) => cmd_ls::run(&cmd.global, &cmd.args),
         #[cfg(all(feature = "fuse", target_os = "linux"))]
         Command::Mount(cmd) => cmd_mount::run(&cmd.global, &cmd.args),
+        Command::RebuildIndex(cmd) => cmd_rebuild_index::run(&cmd.global, &cmd.args),
         Command::Recall(cmd) => cmd_recall::run(&cmd.global, &cmd.args),
         Command::Rechunk(cmd) => cmd_rechunk::run(&cmd.global, &cmd.args),
         Command::Restore(cmd) => cmd_restore::run(&cmd.global, &cmd.args),
@@ -271,6 +274,7 @@ fn extract_global(command: &Command) -> Option<&GlobalArgs> {
         Ls,
         #[cfg(all(feature = "fuse", target_os = "linux"))]
         Mount,
+        RebuildIndex,
         Recall,
         Rechunk,
         Restore,